Add location formatting to user API

+++ pkg/location/coordinates.go
@@ -2,6 +2,6 @@ package location
 
 // Coordinates represents a latitude and longitude representation of a location.
 type Coordinates struct {
-	Latitude  float32 `json:"lat"`
-	Longitude float32 `json:"long"`
+	Latitude  float64 `json:"lat"`
+	Longitude float64 `json:"long"`
 }
